rename Fixfrequencies to NormalizeFrequencies
