+++ common/test.js
@@ -20,7 +20,7 @@ const config = {
 	removeNonASCII: true,
 	compressPlainText: true,
 	encrypt: true,
-	fixFrequencies: true,
+	normalizeFrequencies: true,
 	generateFakeText: true,
 	mapThroughDictionary: false,
 };
