+++ common/transformations/index.js
@@ -1,7 +1,7 @@
 module.exports.RemoveNonASCII = require('./remove-non-ascii');
 module.exports.Compress = require('./compress');
 module.exports.Encrypt = require('./encrypt');
-module.exports.FixFrequencies = require('./fix-frequencies');
+module.exports.NormalizeFrequencies = require('./normalize-frequencies');
 module.exports.GenerateFakeText = require('./generate-fake-text');
 module.exports.GenerateFakeText2 = require('./generate-fake-text-2');
 module.exports.MapThroughDictionary = require('./map-through-dictionary');
@@ -12,7 +12,7 @@ module.exports.createTransformation = (config) => {
     if (config.compressPlainText === true) { transformations.push(module.exports.Compress); }
     // TODO: Assert if password is provided.
     if (config.encrypt === true) { transformations.push(module.exports.Encrypt); }
-    if (config.fixFrequencies === true) { transformations.push(module.exports.FixFrequencies); }
+    if (config.normalizeFrequencies === true) { transformations.push(module.exports.NormalizeFrequencies); }
     if (config.generateFakeText === true) { transformations.push(module.exports.GenerateFakeText2); }
     if (config.mapThroughDictionary === true) { transformations.push(module.exports.MapThroughDictionary); }
