+++ common/transformations/normalize-frequencies.js
